Extend DeviceController interface to include Destroyable
