+++ docs/classes/defaultaudiovideofacade.html
@@ -191,7 +191,7 @@
 					<a name="constructor" class="tsd-anchor"></a>
 					<h3>constructor</h3>
 					<ul class="tsd-signatures tsd-kind-constructor tsd-parent-kind-class">
-						<li class="tsd-signature tsd-kind-icon">new <wbr>Default<wbr>Audio<wbr>Video<wbr>Facade<span class="tsd-signature-symbol">(</span>audioVideoController<span class="tsd-signature-symbol">: </span><a href="../interfaces/audiovideocontroller.html" class="tsd-signature-type">AudioVideoController</a>, videoTileController<span class="tsd-signature-symbol">: </span><a href="../interfaces/videotilecontroller.html" class="tsd-signature-type">VideoTileController</a>, realtimeController<span class="tsd-signature-symbol">: </span><a href="../interfaces/realtimecontroller.html" class="tsd-signature-type">RealtimeController</a>, audioMixController<span class="tsd-signature-symbol">: </span><a href="../interfaces/audiomixcontroller.html" class="tsd-signature-type">AudioMixController</a>, deviceController<span class="tsd-signature-symbol">: </span><a href="../interfaces/devicecontroller.html" class="tsd-signature-type">DeviceController</a>, contentShareController<span class="tsd-signature-symbol">: </span><a href="../interfaces/contentsharecontroller.html" class="tsd-signature-type">ContentShareController</a><span class="tsd-signature-symbol">)</span><span class="tsd-signature-symbol">: </span><a href="defaultaudiovideofacade.html" class="tsd-signature-type">DefaultAudioVideoFacade</a></li>
+						<li class="tsd-signature tsd-kind-icon">new <wbr>Default<wbr>Audio<wbr>Video<wbr>Facade<span class="tsd-signature-symbol">(</span>audioVideoController<span class="tsd-signature-symbol">: </span><a href="../interfaces/audiovideocontroller.html" class="tsd-signature-type">AudioVideoController</a>, videoTileController<span class="tsd-signature-symbol">: </span><a href="../interfaces/videotilecontroller.html" class="tsd-signature-type">VideoTileController</a>, realtimeController<span class="tsd-signature-symbol">: </span><a href="../interfaces/realtimecontroller.html" class="tsd-signature-type">RealtimeController</a>, audioMixController<span class="tsd-signature-symbol">: </span><a href="../interfaces/audiomixcontroller.html" class="tsd-signature-type">AudioMixController</a>, deviceController<span class="tsd-signature-symbol">: </span><span class="tsd-signature-type">DeviceController</span>, contentShareController<span class="tsd-signature-symbol">: </span><a href="../interfaces/contentsharecontroller.html" class="tsd-signature-type">ContentShareController</a><span class="tsd-signature-symbol">)</span><span class="tsd-signature-symbol">: </span><a href="defaultaudiovideofacade.html" class="tsd-signature-type">DefaultAudioVideoFacade</a></li>
 					</ul>
 					<ul class="tsd-descriptions">
 						<li class="tsd-description">
@@ -215,7 +215,7 @@
 									<h5>audioMixController: <a href="../interfaces/audiomixcontroller.html" class="tsd-signature-type">AudioMixController</a></h5>
 								</li>
 								<li>
-									<h5>deviceController: <a href="../interfaces/devicecontroller.html" class="tsd-signature-type">DeviceController</a></h5>
+									<h5>deviceController: <span class="tsd-signature-type">DeviceController</span></h5>
 								</li>
 								<li>
 									<h5>contentShareController: <a href="../interfaces/contentsharecontroller.html" class="tsd-signature-type">ContentShareController</a></h5>
