+++ docs/classes/defaultmeetingsession.html
@@ -203,7 +203,7 @@
 					<a name="devicecontroller" class="tsd-anchor"></a>
 					<h3>device<wbr>Controller</h3>
 					<ul class="tsd-signatures tsd-kind-get-signature tsd-parent-kind-class">
-						<li class="tsd-signature tsd-kind-icon"><span class="tsd-signature-symbol">get</span> deviceController<span class="tsd-signature-symbol">(</span><span class="tsd-signature-symbol">)</span><span class="tsd-signature-symbol">: </span><a href="../interfaces/devicecontroller.html" class="tsd-signature-type">DeviceController</a></li>
+						<li class="tsd-signature tsd-kind-icon"><span class="tsd-signature-symbol">get</span> deviceController<span class="tsd-signature-symbol">(</span><span class="tsd-signature-symbol">)</span><span class="tsd-signature-symbol">: </span><span class="tsd-signature-type">DeviceController</span></li>
 					</ul>
 					<ul class="tsd-descriptions">
 						<li class="tsd-description">
@@ -212,7 +212,7 @@
 									<li>Defined in <a href="https://github.com/aws/amazon-chime-sdk-js/blob/main/src/meetingsession/DefaultMeetingSession.ts#L118">src/meetingsession/DefaultMeetingSession.ts:118</a></li>
 								</ul>
 							</aside>
-							<h4 class="tsd-returns-title">Returns <a href="../interfaces/devicecontroller.html" class="tsd-signature-type">DeviceController</a></h4>
+							<h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">DeviceController</span></h4>
 						</li>
 					</ul>
 				</section>
