+++ docs/interfaces/destroyable.html
@@ -80,6 +80,9 @@
 						<span class="target">Destroyable</span>
 						<ul class="tsd-hierarchy">
 							<li>
+								<a href="devicecontroller.html" class="tsd-signature-type">DeviceController</a>
+							</li>
+							<li>
 								<a href="videoframeprocessor.html" class="tsd-signature-type">VideoFrameProcessor</a>
 							</li>
 						</ul>
