+++ docs/interfaces/meetingsession.html
@@ -133,7 +133,7 @@
 				<section class="tsd-panel tsd-member tsd-kind-property tsd-parent-kind-interface">
 					<a name="devicecontroller" class="tsd-anchor"></a>
 					<h3><span class="tsd-flag ts-flagReadonly">Readonly</span> device<wbr>Controller</h3>
-					<div class="tsd-signature tsd-kind-icon">device<wbr>Controller<span class="tsd-signature-symbol">:</span> <a href="devicecontroller.html" class="tsd-signature-type">DeviceController</a></div>
+					<div class="tsd-signature tsd-kind-icon">device<wbr>Controller<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">DeviceController</span></div>
 					<aside class="tsd-sources">
 						<ul>
 							<li>Defined in <a href="https://github.com/aws/amazon-chime-sdk-js/blob/main/src/meetingsession/MeetingSession.ts#L16">src/meetingsession/MeetingSession.ts:16</a></li>
